Cherry pick JniThreadBinding.h
