+++ platforms/android/tangram/src/main/cpp/jniThreadBinding.h
@@ -1,0 +1,26 @@
+#pragma once
+
+#include <jni.h>
+
+class JniThreadBinding {
+private:
+    JavaVM* jvm;
+    JNIEnv *jniEnv;
+    int status;
+public:
+    JniThreadBinding(JavaVM* _jvm) : jvm(_jvm) {
+        status = jvm->GetEnv((void**)&jniEnv, JNI_VERSION_1_6);
+        if (status == JNI_EDETACHED) { jvm->AttachCurrentThread(&jniEnv, NULL);}
+    }
+    ~JniThreadBinding() {
+        if (status == JNI_EDETACHED) { jvm->DetachCurrentThread(); }
+    }
+
+    JNIEnv* operator->() const {
+        return jniEnv;
+    }
+
+    operator JNIEnv*() const {
+        return jniEnv;
+    }
+};
